pythonCaption/requirements.txt
fastapi
uvicorn
transformers>=4.30.0
Pillow
googletrans==4.0.0-rc1
python-multipart
sentencepiece
protobuf